+++ src/common/file_types.cpp
@@ -38,7 +38,7 @@ file_type_t::get_supported() {
   s_supported_file_types.push_back(file_type_t(Y("FLV (Flash Video)"),                   "flv"));
   s_supported_file_types.push_back(file_type_t(Y("HDMV TextST"),                         "textst"));
   s_supported_file_types.push_back(file_type_t(Y("HEVC/h.265 elementary streams"),       "265 hevc h265 x265"));
-  s_supported_file_types.push_back(file_type_t(Y("IVF with VP8 video files"),            "ivf"));
+  s_supported_file_types.push_back(file_type_t(Y("IVF (AV1, VP8, VP9)"),                 "ivf"));
   s_supported_file_types.push_back(file_type_t(Y("MP4 audio/video files"),               "mp4 m4v"));
   s_supported_file_types.push_back(file_type_t(Y("MPEG audio files"),                    "mp2 mp3"));
   s_supported_file_types.push_back(file_type_t(Y("MPEG program streams"),                "mpg mpeg m2v mpv evo evob vob"));
@@ -84,7 +84,7 @@ file_type_t::get_name(file_type_e type) {
        : file_type_e::hdmv_textst == type ? YT("HDMV TextST subtitles")
        : file_type_e::hevc_es     == type ? YT("HEVC/h.265")
        : file_type_e::hdsub       == type ? YT("HD-DVD subtitles")
-       : file_type_e::ivf         == type ? YT("IVF (VP8/VP9)")
+       : file_type_e::ivf         == type ? YT("IVF (AV1, VP8, VP9)")
        : file_type_e::matroska    == type ? YT("Matroska")
        : file_type_e::microdvd    == type ? YT("MicroDVD")
        : file_type_e::mp3         == type ? YT("MP2/MP3")